Split typedefs.h
