+++ src/main.cpp
@@ -19,6 +19,7 @@ All rights reserved (see LICENSE).
 
 #include "./problems/vrp.h"
 #include "./structures/typedefs.h"
+#include "./structures/cl_args.h"
 #include "./structures/vroom/input/input.h"
 #include "./utils/input_parser.h"
 #include "./utils/output_json.h"
