+++ src/structures/cl_args.h
@@ -1,0 +1,27 @@
+
+#include <string>
+#include <boost/log/trivial.hpp>
+
+struct cl_args_t {
+  // Listing command-line options.
+  std::string osrm_address;                      // -a
+  bool geometry;                                 // -g
+  std::string input_file;                        // -i
+  std::string output_file;                       // -o
+  std::string osrm_port;                         // -p
+  bool use_libosrm;                              // -l
+  boost::log::trivial::severity_level log_level; // -v
+  std::string input;                             // cl arg
+  unsigned nb_threads;                           // -t
+  std::string osrm_profile;                      // -m
+  // Default values.
+  cl_args_t()
+    : osrm_address("0.0.0.0"),
+      geometry(false),
+      osrm_port("5000"),
+      use_libosrm(false),
+      log_level(boost::log::trivial::error),
+      nb_threads(2),
+      osrm_profile("car") {
+  }
+};
