+++ src/structures/typedefs.h
@@ -12,11 +12,9 @@ All rights reserved (see LICENSE).
 
 #include <limits>
 #include <list>
-#include <string>
 #include <unordered_set>
 #include <vector>
 
-#include <boost/log/trivial.hpp>
 #include <boost/optional.hpp>
 
 // To easily differentiate variable types.
@@ -37,30 +35,6 @@ using raw_solution = std::vector<std::list<index_t>>;
 
 // Setting max value would cause trouble with further additions.
 constexpr cost_t INFINITE_COST = 3 * (std::numeric_limits<cost_t>::max() / 4);
-
-struct cl_args_t {
-  // Listing command-line options.
-  std::string osrm_address;                      // -a
-  bool geometry;                                 // -g
-  std::string input_file;                        // -i
-  std::string output_file;                       // -o
-  std::string osrm_port;                         // -p
-  bool use_libosrm;                              // -l
-  boost::log::trivial::severity_level log_level; // -v
-  std::string input;                             // cl arg
-  unsigned nb_threads;                           // -t
-  std::string osrm_profile;                      // -m
-  // Default values.
-  cl_args_t()
-    : osrm_address("0.0.0.0"),
-      geometry(false),
-      osrm_port("5000"),
-      use_libosrm(false),
-      log_level(boost::log::trivial::error),
-      nb_threads(2),
-      osrm_profile("car") {
-  }
-};
 
 // Available location status.
 enum class TYPE { START, JOB, END };
