+++ src/utils/input_parser.h
@@ -13,6 +13,8 @@ All rights reserved (see LICENSE).
 #include "../structures/typedefs.h"
 #include "../structures/vroom/input/input.h"
 
+struct cl_args_t;
+
 input parse(const cl_args_t& cl_args);
 
 #endif
